requirements.txt
py-clob-client>=0.28.0
python-dotenv>=1.0.0
requests>=2.28.0
tabulate>=0.9.0
pandas>=1.5.0
rich>=13.0.0